Add compaction
* If you don't provide the --compact command, nothing will happen
* If you don't provide a compact config for your bundle, nothing will happen
* If you set compact to "images", the default image compact functions
will be used.
* Otherwise, it will use whatever is provided

compact needs to be an object that maps file extensions to functions.
These functions take a buffer and return either a buffer or a Promise
yielding a buffer. All packages from the imagemin project follow this
signature.

Also adjusted the .travis.yml to download the example images and install
the required APTs to compile the C dependencies. Also adjusted the
editorconfig, because YAML files apparently are only allowed to contain
spaces, not tabs.